+++ problems/inc/AoC/2015/problem_04.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
 
 namespace AoC_2015
@@ -7,8 +8,8 @@ namespace AoC_2015
 
 namespace problem_04
 {
-int solve_1( std::istream& input );
-int solve_2( std::istream& input );
+size_t solve_1( std::istream& input );
+size_t solve_2( std::istream& input );
 }  // namespace problem_04
 
 }  // namespace AoC_2015
