+++ problems/inc/AoC/2017/problem_01.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
 
 namespace AoC_2017
@@ -7,8 +8,8 @@ namespace AoC_2017
 
 namespace problem_01
 {
-int solve_1( std::istream& input );
-int solve_2( std::istream& input );
+size_t solve_1( std::istream& input );
+size_t solve_2( std::istream& input );
 }  // namespace problem_01
 
 }  // namespace AoC_2017
