+++ problems/inc/AoC/2017/problem_07.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
 #include <string>
 
@@ -9,7 +10,7 @@ namespace AoC_2017
 namespace problem_07
 {
 std::string solve_1( std::istream& input );
-int solve_2( std::istream& input );
+size_t solve_2( std::istream& input );
 }  // namespace problem_07
 
 }  // namespace AoC_2017
