+++ problems/inc/AoC/2018/problem_02.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
 #include <string>
 
@@ -8,7 +9,7 @@ namespace AoC_2018
 
 namespace problem_02
 {
-int solve_1( std::istream& input );
+size_t solve_1( std::istream& input );
 std::string solve_2( std::istream& input );
 }  // namespace problem_02
