+++ problems/src/2015/problem_01.cpp
@@ -35,7 +35,7 @@ int parse_instruction( const Instruction instruction )
 
 bool is_not_space( const char c )
 {
-  return std::isspace( boost::numeric_cast<int>( c ) ) == 0;
+  return std::isspace( c ) == 0;
 }
 
 auto get_instructions( std::istream& input )
@@ -48,12 +48,12 @@ int calc_last_visited_floor( std::istream& input )
   return ranges::accumulate( get_instructions( input ), 0 );
 }
 
-int calc_number_of_steps_to_reach_basement( std::istream& input )
+size_t calc_number_of_steps_to_reach_basement( std::istream& input )
 {
   auto instructions_to_reach_basement =
       get_instructions( input ) | ranges::view::partial_sum | ranges::view::take_while( []( int floor ) { return floor >= 0; } );
 
-  return 1 + boost::numeric_cast<int>( ranges::distance( instructions_to_reach_basement ) );
+  return 1 + boost::numeric_cast<size_t>( ranges::distance( instructions_to_reach_basement ) );
 }
 
 }  // namespace
@@ -69,7 +69,7 @@ int solve_1( std::istream& input )
   return calc_last_visited_floor( input );
 }
 
-int solve_2( std::istream& input )
+size_t solve_2( std::istream& input )
 {
   return calc_number_of_steps_to_reach_basement( input );
 }
