+++ problems/src/2015/problem_08.cpp
@@ -8,7 +8,6 @@
 #include "range/v3/view/getlines.hpp"
 #include "range/v3/view/transform.hpp"
 
-#include "boost/numeric/conversion/cast.hpp"
 #include "boost/spirit/home/x3.hpp"
 
 #include <istream>
@@ -93,14 +92,14 @@ namespace AoC_2015
 namespace problem_08
 {
 
-int solve_1( std::istream& input )
+size_t solve_1( std::istream& input )
 {
-  return boost::numeric_cast<int>( calc_lines<&calc_encoded_line_len_1>( input ) );
+  return calc_lines<&calc_encoded_line_len_1>( input );
 }
 
-int solve_2( std::istream& input )
+size_t solve_2( std::istream& input )
 {
-  return boost::numeric_cast<int>( calc_lines<&calc_encoded_line_len_2>( input ) );
+  return calc_lines<&calc_encoded_line_len_2>( input );
 }
 
 AOC_REGISTER_PROBLEM( 2015_08, solve_1, solve_2 );
